requirements.txt
apturl==0.5.2
asgiref==3.5.2
asn1crypto==0.24.0
attrs==17.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.1.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro-info===0.18ubuntu0.18.04.1
dj-database-url==1.0.0
Django==4.1.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.7.1
django-language-field==0.0.3
django-model-utils==4.1.1
django-rest-framework==0.1.0
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
import-export==0.2.67.dev6
joblib==0.11
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==2.1.1
netifaces==0.10.4
nose==1.3.7
numexpr==2.6.4
numpy==1.13.3
oauth==1.0.1
oauthlib==3.0.1
odfpy==1.4.1
olefile==0.45.1
openpyxl==3.0.10
pandas==0.22.0
pep8==1.7.1
pexpect==4.2.1
Pillow==6.2.2
pluggy==0.6.0
protobuf==3.0.0
psycopg2==2.9.3
py==1.5.2
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pytest==3.3.2
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.5
pyxdg==0.25
PyYAML==6.0
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
scikit-learn==0.19.1
scipy==0.19.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
ssh-import-id==5.7
stripe==2.27.0
system-service==0.3
systemd-python==234
tables==3.4.2
tablib==3.2.1
twilio==7.3.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
whitenoise==6.2.0
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==4.3.2
